Separate out algorithm panel, parameter panel, and graph panel.
