+++ tetrad-gui/src/main/java/edu/cmu/tetradapp/editor/algorithm/GraphCard.java
@@ -1,0 +1,218 @@
+/*
+ * Copyright (C) 2019 University of Pittsburgh.
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
+ * MA 02110-1301  USA
+ */
+package edu.cmu.tetradapp.editor.algorithm;
+
+import edu.cmu.tetrad.graph.Graph;
+import edu.cmu.tetradapp.editor.GeneralAlgorithmEditor;
+import edu.cmu.tetradapp.model.GeneralAlgorithmRunner;
+import edu.cmu.tetradapp.ui.PaddingPanel;
+import edu.cmu.tetradapp.util.BootstrapTable;
+import edu.cmu.tetradapp.util.ImageUtils;
+import edu.cmu.tetradapp.workbench.GraphWorkbench;
+import java.awt.BorderLayout;
+import java.awt.Dimension;
+import java.awt.Font;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+import java.net.URL;
+import javax.help.CSH;
+import javax.help.HelpBroker;
+import javax.help.HelpSet;
+import javax.swing.Box;
+import javax.swing.GroupLayout;
+import javax.swing.ImageIcon;
+import javax.swing.JButton;
+import javax.swing.JLabel;
+import javax.swing.JPanel;
+import javax.swing.JScrollPane;
+import javax.swing.JSplitPane;
+import javax.swing.JTabbedPane;
+import javax.swing.border.EmptyBorder;
+
+/**
+ *
+ * Apr 15, 2019 4:49:15 PM
+ *
+ * @author Kevin V. Bui (kvb2@pitt.edu)
+ */
+public class GraphCard extends JPanel {
+
+    private static final long serialVersionUID = 8349434413076189088L;
+
+    private final JButton backBtn = new JButton("<   Set Parameters");
+
+    private final JLabel title = new JLabel();
+    private final JTabbedPane graphContainer = new JTabbedPane();
+
+    private final GeneralAlgorithmRunner algorithmRunner;
+
+    public GraphCard(GeneralAlgorithmEditor algorithmEditor, GeneralAlgorithmRunner algorithmRunner) {
+        this.algorithmRunner = algorithmRunner;
+
+        initComponents(algorithmEditor);
+    }
+
+    private void initComponents(GeneralAlgorithmEditor algorithmEditor) {
+        Dimension buttonSize = new Dimension(268, 25);
+        backBtn.setMinimumSize(buttonSize);
+        backBtn.setMaximumSize(buttonSize);
+        backBtn.addActionListener((e) -> {
+            firePropertyChange("graphBack", null, null);
+        });
+
+        JPanel titlePanel = new JPanel();
+        titlePanel.add(title);
+
+        add(titlePanel, BorderLayout.NORTH);
+        add(new PaddingPanel(graphContainer), BorderLayout.CENTER);
+        add(new SouthPanel(), BorderLayout.SOUTH);
+
+        addPropertyChangeListener(algorithmEditor);
+    }
+
+    /**
+     * Resulting graph with bootstrap table - Zhou
+     *
+     * @param graph
+     * @return
+     */
+    private JSplitPane createSearchResultPane(Graph graph) {
+        // topBox contains the graphEditorScroll and the instructionBox underneath
+        Box topBox = Box.createVerticalBox();
+        topBox.setPreferredSize(new Dimension(820, 400));
+
+        GraphWorkbench graphWorkbench = new GraphWorkbench(graph);
+        graphWorkbench.enableEditing(false);
+
+        // topBox graph editor
+        JScrollPane graphEditorScroll = new JScrollPane();
+        graphEditorScroll.setPreferredSize(new Dimension(820, 420));
+        graphEditorScroll.setViewportView(graphWorkbench);
+
+        // Instruction with info button
+        Box instructionBox = Box.createHorizontalBox();
+        instructionBox.setMaximumSize(new Dimension(820, 40));
+
+        JLabel label = new JLabel("More information on graph edge types");
+        label.setFont(new Font("SansSerif", Font.PLAIN, 12));
+
+        // Info button added by Zhou to show edge types
+        JButton infoBtn = new JButton(new ImageIcon(ImageUtils.getImage(this, "info.png")));
+        infoBtn.setBorder(new EmptyBorder(0, 0, 0, 0));
+
+        // Clock info button to show edge types instructions - Zhou
+        infoBtn.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                // Initialize helpSet
+                String helpHS = "/resources/javahelp/TetradHelp.hs";
+
+                try {
+                    URL url = this.getClass().getResource(helpHS);
+                    HelpSet helpSet = new HelpSet(null, url);
+
+                    helpSet.setHomeID("graph_edge_types");
+                    HelpBroker broker = helpSet.createHelpBroker();
+                    ActionListener listener = new CSH.DisplayHelpFromSource(broker);
+                    listener.actionPerformed(e);
+                } catch (Exception ee) {
+                    System.out.println("HelpSet " + ee.getMessage());
+                    System.out.println("HelpSet " + helpHS + " not found");
+                    throw new IllegalArgumentException();
+                }
+            }
+        });
+
+        instructionBox.add(label);
+        instructionBox.add(Box.createHorizontalStrut(2));
+        instructionBox.add(infoBtn);
+
+        // Add to topBox
+        topBox.add(graphEditorScroll);
+        topBox.add(instructionBox);
+
+        // bottomBox contains bootstrap table
+        Box bottomBox = Box.createVerticalBox();
+        bottomBox.setPreferredSize(new Dimension(820, 150));
+
+        bottomBox.add(Box.createVerticalStrut(5));
+
+        // Put the table title label in a box so it can be centered
+        Box tableTitleBox = Box.createHorizontalBox();
+        JLabel tableTitle = new JLabel("Edges and Edge Type Probabilities");
+        tableTitleBox.add(tableTitle);
+
+        bottomBox.add(tableTitleBox);
+
+        bottomBox.add(Box.createVerticalStrut(5));
+
+        JScrollPane tablePane = BootstrapTable.renderBootstrapTable(graph);
+
+        bottomBox.add(tablePane);
+
+        // Use JSplitPane to allow resize the bottom box - Zhou
+        JSplitPane splitPane = new JSplitPane(JSplitPane.VERTICAL_SPLIT, topBox, bottomBox);
+        splitPane.setDividerLocation(400);
+
+        return splitPane;
+    }
+
+    public void refresh() {
+        graphContainer.removeAll();
+
+        title.setText("Algorithm: " + algorithmRunner.getAlgorithmName());
+
+        int count = 0;
+        for (Graph graph : algorithmRunner.getGraphs()) {
+            graphContainer.addTab(
+                    String.format("Graph %d", ++count),
+                    new PaddingPanel(createSearchResultPane(graph)));
+        }
+    }
+
+    private class SouthPanel extends JPanel {
+
+        private static final long serialVersionUID = -6938352710872851817L;
+
+        public SouthPanel() {
+            initComponents();
+        }
+
+        private void initComponents() {
+            GroupLayout layout = new GroupLayout(this);
+            layout.setHorizontalGroup(
+                    layout.createParallelGroup(GroupLayout.Alignment.LEADING)
+                            .addGroup(layout.createSequentialGroup()
+                                    .addContainerGap(GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                    .addComponent(backBtn)
+                                    .addContainerGap(GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+            );
+            layout.setVerticalGroup(
+                    layout.createParallelGroup(GroupLayout.Alignment.LEADING)
+                            .addGroup(layout.createSequentialGroup()
+                                    .addContainerGap()
+                                    .addComponent(backBtn)
+                                    .addContainerGap())
+            );
+
+            this.setLayout(layout);
+        }
+    }
+
+}
